requirements.txt
flask>=3.0.0
Pillow>=10.0.0
pillow-heif>=0.15.0
pypdf>=3.0.0
PyMuPDF>=1.23.0
easyocr>=1.7.0
pytesseract>=0.3.10
requests>=2.31.0
yt-dlp>=2024.1.0
